Add medication and dosage regimens to the JSON files and update the MedicationEntity schema

+++ shared/src/commonMain/kotlin/org/dishch/medcalculator/data/local/DatabaseInitializer.kt
@@ -14,8 +14,8 @@ import kotlinx.coroutines.withContext
 data class MedicationJson(
     val id: Long,
     val name: String,
-    val dosageMgPerMl: Int,
-    val maxSingleDoseMg: Int,
+    val dosageMgPerMl: Double,
+    val maxSingleDoseMg: Double,
     val minAgeMonths: Int = 0
 ) {
     fun toEntity() = MedicationEntity(
